Module spirv_std_types::image_params

Enums

AccessQualifier

The access permissions for the image.

Arrayed

Whether the image uses arrayed content.

Dimensionality

The dimension of the image.

ImageDepth

Whether a given image contains depth information. Note Whether or not to perform depth comparisons is a property of the sampling code, not of this type.

ImageFormat

The underlying internal representation of the image.

Multisampled

Whether the image uses arrayed content.

Sampled

Whether or not the image will be accessed in combination with a sampler.